- add .vscode to .gitignore - move generic methods to same class - use pytest
