+++ .gitignore
@@ -4,3 +4,4 @@ __pycache__
 dist
 *.egg-info
 venv
+.vscode
